Parse labels by changing grammar
I store the current and previous tokens. It takes a lot more work to add
parsing of labels, where we may be in an expression or a label, in the
`parseStmt` function. It's possible by "properly" doing it, that is
mostly matching on the "previous" token, but I'd rather slightly change
the grammar to make parsing labels easy. The easier it is to parse the
better time we'll all have so it's a fair trade off to me.