Add function call args and return values
There are a few things wrong here, like the string maps are really
inefficient, most likely that will be replaced with some function header
in the bytecode. But, this is not meant to be efficient yet, so it's
fine. Also return values are pushed even if the return type is void for
two reasons: One so that all value statements can just pop a value off
the stack, and two because we don't worry about type checking or
resolution or anything yet. This should be okay for now, but in the
future it may be worth it to go through these cases and tidy it up.