Use new `PassManager` in tests
This isn't a ton easier yet, but now you can run a bunch of passes
simply when there are more optimizations. Unfortunately can't get the
return type of the pass properly because the program isn't comptime.
Another solution for that problem could be to pass in a comptime return
type for other passes.