[WIP] Move towards some parser improvements
This is in the spirit of using a rules array for the Pratt parsing rules
to make adding stuff easier. To do this reasonably, it also requires
some other changes, like keywords being a separate field and function
calls storing a Value as its function access.

Given that, this is a partway that doesn't work, but does compile.